Prefactor: Move `seedidentity` pkg to `gardenletidentity/seed`
In the next commit, we introduce a similar package for gardenlet shoot
identities.
